requirements.txt
xlsxwriter==3.0.9
requests==2.28.2
bs4==0.0.1
mailjet_rest==1.3.4
python-dotenv==1.0.0
pandas~=1.5.1